requirements/production.txt
alabaster==0.7.9
anyjson==0.3.3
Babel==2.3.4
defusedxml==0.4.1
Django==1.10.1
django-allauth==0.27.0
django-appconf==1.0.2
django-compressor==2.1

django-su==0.5.2
djangorestframework==3.4.6
docutils==0.12
ecdsa==0.13
Fabric==1.12.0
gunicorn==19.6.0
imagesize==0.7.1
Jinja2==2.8
MarkupSafe==0.23

nose==1.3.7
numpy==1.11.1
oauthlib==2.0.0
pandas==0.18.1
paramiko==1.17.2
psycopg2==2.6.2
pycrypto==2.6.1
Pygments==2.1.3
python-dateutil==2.5.3
python3-openid==3.0.10
pytz==2016.6.1
PyYAML==3.12
rcssmin==1.0.6
requests==2.11.1
requests-oauthlib==0.6.2
rest-pandas==0.4.1
rjsmin==1.0.12
six==1.10.0
snowballstemmer==1.2.1

voluptuous==0.9.3
